// admin/src/auth/types.ts
export interface UserRegisterRequest {
  username: string;
  phone_number: string;
  email: string;
  password: string;
  citizenship_no: string;
  document_type: DocumentType;
  documents: File[]; // <-- added
}

export enum DocumentType {
  CITIZENSHIP = "CITIZENSHIP",
  NATIONAL_ID = "NATIONAL_ID",
  PASSPORT = "PASSPORT",
}


export interface UserLoginRequest {
  email: string;
  password: string;
}



export type UserLoginResponse = {
    created_at: Date | null;
    id: string;
    username: string;
    phone_number: string;
    email: string | null;
    password_hash: string;
    user_type: "voter"|"admin" | null;
    is_verified: boolean | null;
    citizenship_no: string | null;
}

export interface Party {
  id: string;
  party_name: string;
  abbreviation: string;
  leader_name: string;
  is_verified: boolean;
  created_at: string;
  registered_by: string | null;
  party_documents: any[]; 
}